backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
pydantic==2.7.4
pydantic-settings==2.3.4
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
